adding photo management

+++ mazad_service/src/main/java/ch/com/mazad/repository/AddressRepository.java
@@ -1,10 +1,10 @@
 package ch.com.mazad.repository;
 
 import ch.com.mazad.domain.Address;
-import ch.com.mazad.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -15,5 +15,5 @@ public interface AddressRepository extends JpaRepository<Address, Long>
     Optional<Address> findOneByReference(String ref);
 
     @Query(value = "select ad.id from address ad where ad.reference=?1",nativeQuery = true)
-    Optional<Long> findIdByReference(String ref);
+    Optional<BigInteger> findIdByReference(String ref);
 }
