+++ mazad_service/src/main/java/ch/com/mazad/repository/SaleRepository.java
@@ -1,10 +1,10 @@
 package ch.com.mazad.repository;
 
-import ch.com.mazad.domain.Bid;
 import ch.com.mazad.domain.Sale;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -15,5 +15,5 @@ public interface SaleRepository extends JpaRepository<Sale,Long> {
     Optional<Sale> findOneByReference(String ref);
 
     @Query(value = "select b.id from sale b where b.reference=?1",nativeQuery = true)
-    Optional<Long> findIdByReference(String ref);
+    Optional<BigInteger> findIdByReference(String ref);
 }
