+++ mazad_service/src/main/java/ch/com/mazad/repository/UserRepository.java
@@ -4,6 +4,7 @@ import ch.com.mazad.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -23,6 +24,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<String> getUserReferenceByEmail(String email);
 
     @Query(value = "select us.id from user us where b.reference=?1", nativeQuery = true)
-    Optional<Long> findIdByReference(String ref);
+    Optional<BigInteger> findIdByReference(String ref);
 
 }
